requirements.txt
ipykernel
oci
python-dotenv
uuid
